# openapi-chatgpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crowley ChatGPT Actions API",
    "version": "3.9.15",
    "description": "GPT Toolbelt gateway (V3.9.15). Primary surface: GET /catalog, POST /read, POST /write with tool names from the catalog. Legacy per-route paths remain as deprecated aliases for one release. Bearer auth required. Do not call full /api/*."
  },
  "servers": [
    {
      "url": "https://YOUR-CROWLEY-DOMAIN",
      "description": "Public HTTPS endpoint forwarding to local Crowley on 127.0.0.1:8765"
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/actions/health": {
      "get": {
        "operationId": "actionsHealth",
        "summary": "Verify bearer auth and runtime",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LooseObject" }
              }
            }
          }
        }
      }
    },
    "/api/actions/catalog": {
      "get": {
        "operationId": "actionsCatalog",
        "summary": "List registered tools and arg schemas",
        "description": "Call after health check to discover tool names for POST /read and POST /write.",
        "responses": {
          "200": {
            "description": "Tool catalog",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LooseObject" }
              }
            }
          }
        }
      }
    },
    "/api/actions/read": {
      "post": {
        "operationId": "actionsRead",
        "summary": "Invoke a read tool",
        "description": "Examples: context.get, memory.get, ticket.list, agent.sync, inspect.writeback_result, github.file. See GET /catalog.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ActionsInvokeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LooseObject" }
              }
            }
          }
        }
      }
    },
    "/api/actions/write": {
      "post": {
        "operationId": "actionsWrite",
        "summary": "Invoke a write tool",
        "description": "Examples: writeback.ingest, ticket.create, handoff.ingest. See GET /catalog.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ActionsInvokeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LooseObject" }
              }
            }
          }
        }
      }
    },
    "/api/actions/agent/sync": {
      "get": {
        "operationId": "actionsAgentSync",
        "summary": "Alias route for agent.sync",
        "description": "Convenience alias for read tool `agent.sync`.",
        "parameters": [
          { "name": "agent", "in": "query", "required": false, "schema": { "type": "string", "default": "chatgpt" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Tool result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LooseObject" } } }
          }
        }
      }
    },
    "/api/actions/agent/deep_sync": {
      "get": {
        "operationId": "actionsAgentDeepSync",
        "summary": "Alias route for agent.deep_sync",
        "description": "Convenience alias for read tool `agent.deep_sync`.",
        "parameters": [
          { "name": "section", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "agent", "in": "query", "required": false, "schema": { "type": "string", "default": "chatgpt" } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Tool result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LooseObject" } } }
          }
        }
      }
    },
    "/api/actions/github/status": {
      "get": {
        "operationId": "actionsGithubStatus",
        "summary": "Alias route for github.status",
        "responses": {
          "200": {
            "description": "Tool result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LooseObject" } } }
          }
        }
      }
    },
    "/api/actions/github/file": {
      "get": {
        "operationId": "actionsGithubFile",
        "summary": "Alias route for github.file",
        "parameters": [
          { "name": "path", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "ref", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Tool result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LooseObject" } } }
          }
        }
      }
    },
    "/api/actions/github/search_code": {
      "get": {
        "operationId": "actionsGithubSearchCode",
        "summary": "Alias route for github.search_code",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "ref", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Tool result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LooseObject" } } }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ActionsInvokeRequest": {
        "type": "object",
        "required": ["tool"],
        "properties": {
          "tool": {
            "type": "string",
            "description": "Registered tool name from /catalog (e.g. memory.get, ticket.create)"
          },
          "args": {
            "type": "object",
            "properties": {},
            "additionalProperties": true,
            "description": "Tool-specific arguments object"
          }
        }
      },
      "LooseObject": {
        "type": "object",
        "properties": {},
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  }
}